requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
langchain_community
langchain_core
langchain_huggingface
langchain_openai
flask
python-dotenv
pypdf
-e .
